fix linters

+++ geoproject/services/locations.py
@@ -2,9 +2,12 @@ from sqlalchemy.orm import Session
 from geoproject.models.locations import DBLocations
 from geoproject.config.database import get_session
 from geoproject.models.locations import LocationCreateUpdate, LocationResponse
-from fastapi import Depends
+from fastapi import Depends, HTTPException
 from uuid import UUID
 from geoalchemy2.shape import to_shape
+from geoalchemy2 import WKTElement
+from typing import cast
+from datetime import datetime
 
 
 def create_location_service(
@@ -16,12 +19,11 @@ def create_location_service(
     db.add(db_location)
     db.commit()
     db.refresh(db_location)
-
-    geometry = to_shape(db_location.geometry).__geo_interface__
+    geometry = to_shape(cast(WKTElement, db_location.geometry)).__geo_interface__
     return LocationResponse(
-        location_id=db_location.location_id,
-        timestamp=db_location.timestamp,
-        description=db_location.description,
+        location_id=cast(UUID, db_location.location_id),
+        timestamp=cast(datetime, db_location.timestamp),
+        description=cast(str, db_location.description),
         geometry=geometry,
     )
 
@@ -30,29 +32,37 @@ def get_all_locations_service(db: Session = Depends(get_session)):
     results = db.query(DBLocations).all()
     locations = []
     for result in results:
-        geometry = to_shape(result.geometry).__geo_interface__
+        geometry = to_shape(cast(WKTElement, result.geometry)).__geo_interface__
         locations.append(
             LocationResponse(
-                location_id=result.location_id,
-                timestamp=result.timestamp,
-                description=result.description,
+                location_id=cast(UUID, result.location_id),
+                timestamp=cast(datetime, result.timestamp),
+                description=cast(str, result.description),
                 geometry=geometry,
             )
         )
     return locations
 
 
-def get_location_by_id_service(location_id: UUID, db: Session = Depends(get_session)):
-    result = (
+def get_location_by_id(
+    location_id: UUID, db: Session = Depends(get_session)
+) -> DBLocations | None:
+    return (
         db.query(DBLocations)
         .filter(DBLocations.location_id == str(location_id))
         .first()
     )
-    geometry = to_shape(result.geometry).__geo_interface__
+
+
+def get_location_by_id_service(location_id: UUID, db: Session = Depends(get_session)):
+    result = get_location_by_id(location_id, db)
+    if not result:
+        raise HTTPException(status_code=404, detail="Location not found")
+    geometry = to_shape(cast(WKTElement, result.geometry)).__geo_interface__
     return LocationResponse(
-        location_id=result.location_id,
-        timestamp=result.timestamp,
-        description=result.description,
+        location_id=cast(UUID, result.location_id),
+        timestamp=cast(datetime, result.timestamp),
+        description=cast(str, result.description),
         geometry=geometry,
     )
 
@@ -63,36 +73,32 @@ def update_location_service(
     db: Session = Depends(get_session),
 ):
     location_data = location.model_dump(exclude_unset=True)
-    db_location = (
-        db.query(DBLocations)
-        .filter(DBLocations.location_id == str(location_id))
-        .first()
-    )
+    db_location = get_location_by_id(location_id, db)
+    if not db_location:
+        raise HTTPException(status_code=404, detail="Location not found")
     db_location.description = location_data["description"]
     db_location.geometry = location_data["geometry"]
     db.commit()
     db.refresh(db_location)
-    geometry = to_shape(db_location.geometry).__geo_interface__
+    geometry = to_shape(cast(WKTElement, db_location.geometry)).__geo_interface__
     return LocationResponse(
-        location_id=db_location.location_id,
-        timestamp=db_location.timestamp,
-        description=db_location.description,
+        location_id=cast(UUID, db_location.location_id),
+        timestamp=cast(datetime, db_location.timestamp),
+        description=cast(str, db_location.description),
         geometry=geometry,
     )
 
 
 def delete_location_service(location_id: UUID, db: Session = Depends(get_session)):
-    db_location = (
-        db.query(DBLocations)
-        .filter(DBLocations.location_id == str(location_id))
-        .first()
-    )
+    db_location = get_location_by_id(location_id, db)
+    if not db_location:
+        raise HTTPException(status_code=404, detail="Location not found")
     db.delete(db_location)
     db.commit()
-    geometry = to_shape(db_location.geometry).__geo_interface__
+    geometry = to_shape(cast(WKTElement, db_location.geometry)).__geo_interface__
     return LocationResponse(
-        location_id=db_location.location_id,
-        timestamp=db_location.timestamp,
-        description=db_location.description,
+        location_id=cast(UUID, db_location.location_id),
+        timestamp=cast(datetime, db_location.timestamp),
+        description=cast(str, db_location.description),
         geometry=geometry,
     )
